+++ src/api/returnApi.tsx
@@ -8,7 +8,7 @@ export const returnApi = async (cabinetId: number) => {
       return {
         message: "반납 성공",
         data: response.data,
-        state: response.status,
+        status: response.status,
       };
     }
   } catch (error) {
