+++ src/hooks/Admin/useAdminStatus.tsx
@@ -1,4 +1,4 @@
-import { useEffect, useState } from "react";
+import { useState } from "react";
 import { SelectedCabinet, StatusData } from "@/types/CabinetType";
 import { SelectedMultiCabinetsData } from "@/types/MultiCabinetType";
 import { CabinetStatus, CabinetStatusType } from "@/types/StatusEnum";
